src/test/java/ru/job4j/condition/Task43Test.java
package ru.job4j.condition;

import org.junit.Test;

import java.io.ByteArrayOutputStream;
import java.io.PrintStream;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.*;

public class Task43Test {

    public String ln = System.lineSeparator();

    @Test
    public void when200ThenYes() {
        ByteArrayOutputStream out = new ByteArrayOutputStream();
        System.setOut(new PrintStream(out));
        Task43.containsZero(200);
        String expected = "Yes" + ln;
        assertThat(out.toString(), is(expected));
    }

    @Test
    public void when201ThenYes() {
        ByteArrayOutputStream out = new ByteArrayOutputStream();
        System.setOut(new PrintStream(out));
        Task43.containsZero(201);
        String expected = "Yes" + ln;
        assertThat(out.toString(), is(expected));
    }

    @Test
    public void when221ThenNo() {
        ByteArrayOutputStream out = new ByteArrayOutputStream();
        System.setOut(new PrintStream(out));
        Task43.containsZero(221);
        String expected = "No" + ln;
        assertThat(out.toString(), is(expected));
    }

    @Test
    public void when999ThenNo() {
        ByteArrayOutputStream out = new ByteArrayOutputStream();
        System.setOut(new PrintStream(out));
        Task43.containsZero(999);
        String expected = "No" + ln;
        assertThat(out.toString(), is(expected));
    }
}